TST Reorganized tests. ENH Improvements to atom types.
