+++ test/README.md
@@ -1,0 +1,3 @@
+```
+julia> using Base.Test, JuQ, JuQ._k; include("test/lowlevel-tests.jl");
+```
